Commonly used abbreviations in R code

Returns a tibble listing an opinionated set of abbreviations commonly used when writing R code. It can be used as a reference, for example, to check availability when considering using some abbreviation in a function or argument name.

abbreviations(expand = FALSE)

abbrs(expand = FALSE)

Arguments

expand
Whether to expand the full_expression column and return the data in long format. If FALSE, a "nested" list column full_expressions will be returned, meaning the values in column abbreviation will be unique.

Value

A tibble.

Examples

pkgsnip::abbrs(expand = TRUE) |> print(n = Inf)

#> # A tibble: 138 × 2
#>     full_expression       abbreviation
#>     <chr>                 <chr>       
#>   1 abbreviate            abbr        
#>   2 abbreviation          abbr        
#>   3 abbreviations         abbrs       
#>   4 absolute              abs         
#>   5 argument              arg         
#>   6 arguments             args        
#>   7 attribute             attr        
#>   8 attributes            attrs       
#>   9 authenticate          auth        
#>  10 authentication        auth        
#>  11 authentications       auths       
#>  12 bibliography          bib         
#>  13 bibliographies        bibs        
#>  14 character             chr         
#>  15 characters            chrs        
#>  16 column                col         
#>  17 columns               cols        
#>  18 command               cmd         
#>  19 commands              cmds        
#>  20 combination           combo       
#>  21 combinations          combos      
#>  22 condition             cnd         
#>  23 conditions            cnds        
#>  24 configure             config      
#>  25 configuration         config      
#>  26 configurations        configs     
#>  27 database              db          
#>  28 dataframe             df          
#>  29 dataframes            dfs         
#>  30 dataframe column      dfc         
#>  31 dataframe row         dfr         
#>  32 depend                dep         
#>  33 dependency            dep         
#>  34 dependencies          deps        
#>  35 develop               dev         
#>  36 development           dev         
#>  37 developer             dev         
#>  38 developments          devs        
#>  39 developers            devs        
#>  40 differentiate         diff        
#>  41 difference            diff        
#>  42 differences           diffs       
#>  43 directory             dir         
#>  44 directories           dirs        
#>  45 distribution          distro      
#>  46 distributions         distros     
#>  47 document              doc         
#>  48 documents             docs        
#>  49 double                dbl         
#>  50 doubles               dbls        
#>  51 element               el          
#>  52 elements              els         
#>  53 environment           env         
#>  54 environments          envs        
#>  55 exclude               excl        
#>  56 exclusion             excl        
#>  57 expression            expr        
#>  58 expressions           exprs       
#>  59 factor                fct         
#>  60 factors               fcts        
#>  61 filesystem            fs          
#>  62 formula               fm          
#>  63 formulas              fms         
#>  64 formulae              fms         
#>  65 function              fn          
#>  66 functions             fns         
#>  67 generate              gen         
#>  68 generation            gen         
#>  69 generations           gens        
#>  70 identify              id          
#>  71 identifier            id          
#>  72 identifiers           ids         
#>  73 include               incl        
#>  74 inclusion             incl        
#>  75 index                 i           
#>  76 indices/indexes       ix          
#>  77 information           info        
#>  78 initialize            init        
#>  79 initialization        init        
#>  80 integer               int         
#>  81 integers              ints        
#>  82 label                 lbl         
#>  83 labels                lbls        
#>  84 language              lang        
#>  85 languages             langs       
#>  86 level                 lvl         
#>  87 levels                lvls        
#>  88 list                  ls          
#>  89 logical               lgl         
#>  90 logicals              lgls        
#>  91 Markdown              md          
#>  92 message               msg         
#>  93 messages              msgs        
#>  94 modify                mod         
#>  95 modification          mod         
#>  96 modifications         mods        
#>  97 number                nr          
#>  98 numbers               nrs         
#>  99 number of             n           
#> 100 numeric               num         
#> 101 numerics              nums        
#> 102 object                obj         
#> 103 objects               objs        
#> 104 option                opt         
#> 105 options               opts        
#> 106 package               pkg         
#> 107 packages              pkgs        
#> 108 parameterize          param       
#> 109 parameter             param       
#> 110 parameters            params      
#> 111 R Markdown            rmd         
#> 112 refer                 ref         
#> 113 reference             ref         
#> 114 references            refs        
#> 115 regular expression(s) regex       
#> 116 relative              rel         
#> 117 remove                rm          
#> 118 roxygen2              roxy        
#> 119 separate              sep         
#> 120 separator             sep         
#> 121 separators            seps        
#> 122 sequential            seq         
#> 123 sequence              seq         
#> 124 sequences             seqs        
#> 125 specify               spec        
#> 126 specification         spec        
#> 127 specifications        specs       
#> 128 string                str         
#> 129 strings               strs        
#> 130 temporary             tmp         
#> 131 value                 val         
#> 132 values                vals        
#> 133 variable              v           
#> 134 variables             vx          
#> 135 vectorize             vctr        
#> 136 vector                vctr        
#> 137 vectors               vctrs       
#> 138 working directory     wd