Commonly used abbreviations in R code

Returns a tibble listing an opinionated set of abbreviations commonly used when writing R code. It can be used as a reference, for example, to check availability/consistency when considering using some abbreviation in a function or argument name.

Usage

abbrs(unnest = FALSE)

Arguments

unnest

Whether to unnest the full_expression column and return the data in long format. If FALSE, a "nested" list column full_expressions will be returned, meaning the values in column abbreviation will be unique.

Value

A tibble.

Details

To avoid the list getting excessively long and redundant, the main focus is on verbs and nouns. Adjectives and adverbs are only listed in column full_expression when there's not already a verb or noun with the same root word listed for a particular abbreviation. That means that e.g. the abbreviation "seq" could also stand for the full expressions "sequential" or "sequentially", despite the fact that it is not explicitly listed but merely the verb/noun "sequence" is.

Motivation

To quote Leon Bambrick:

"There are 2 hard problems in computer science: cache invalidation, naming things, and off-by-1 errors."

Examples

pkgsnip::abbrs(unnest = TRUE) |> print(n = Inf)
#> # A tibble: 271 × 2
#>     full_expression     abbreviation
#>     <chr>               <chr>       
#>   1 abbreviate          abbr        
#>   2 abbreviation        abbr        
#>   3 abbreviations       abbrs       
#>   4 absolute            abs         
#>   5 argument            arg         
#>   6 arguments           args        
#>   7 attribute           attr        
#>   8 attributes          attrs       
#>   9 authenticate        auth        
#>  10 authentication      auth        
#>  11 authentications     auths       
#>  12 auxiliary           aux         
#>  13 background          bg          
#>  14 back up             kbp         
#>  15 backup              bkp         
#>  16 bibliography        bib         
#>  17 bibliographies      bibs        
#>  18 certify             cert        
#>  19 certificate         cert        
#>  20 certification       cert        
#>  21 certificates        certs       
#>  22 certifications      certs       
#>  23 chapter             chpt        
#>  24 chapters            chpts       
#>  25 character           chr         
#>  26 characters          chrs        
#>  27 column              col         
#>  28 columns             cols        
#>  29 command             cmd         
#>  30 commands            cmds        
#>  31 combination         combo       
#>  32 combinations        combos      
#>  33 condition           cnd         
#>  34 conditions          cnds        
#>  35 configure           config      
#>  36 configuration       config      
#>  37 configurations      configs     
#>  38 connection          conn        
#>  39 connections         conns       
#>  40 current             cur         
#>  41 database            db          
#>  42 dataframe           df          
#>  43 dataframes          dfs         
#>  44 dataframe column    dfc         
#>  45 dataframe row       dfr         
#>  46 define              def         
#>  47 definition          def         
#>  48 definitions         defs        
#>  49 delete              del         
#>  50 deletion            del         
#>  51 deletions           dels        
#>  52 depend              dep         
#>  53 dependency          dep         
#>  54 dependencies        deps        
#>  55 develop             dev         
#>  56 development         dev         
#>  57 developer           dev         
#>  58 developments        devs        
#>  59 developers          devs        
#>  60 dictionary          dict        
#>  61 dictionaries        dicts       
#>  62 differentiate       diff        
#>  63 difference          diff        
#>  64 differences         diffs       
#>  65 directory           dir         
#>  66 directories         dirs        
#>  67 distribution        distro      
#>  68 distributions       distros     
#>  69 document            doc         
#>  70 documents           docs        
#>  71 double              dbl         
#>  72 doubles             dbls        
#>  73 duplicate           dupl        
#>  74 duplication         dupl        
#>  75 duplicates          dupls       
#>  76 duplications        dupls       
#>  77 dynamic             dyn         
#>  78 element             el          
#>  79 elements            els         
#>  80 enumerate           enum        
#>  81 enumeration         enum        
#>  82 enumerations        enums       
#>  83 environment         env         
#>  84 environments        envs        
#>  85 evaluate            eval        
#>  86 evaluation          eval        
#>  87 evaluations         evals       
#>  88 execute             exec        
#>  89 execution           exec        
#>  90 executions          execs       
#>  91 exclude             excl        
#>  92 exclusion           excl        
#>  93 expression          expr        
#>  94 expressions         exprs       
#>  95 extend              ext         
#>  96 extension           ext         
#>  97 extensions          exts        
#>  98 factor              fct         
#>  99 factors             fcts        
#> 100 figure              fig         
#> 101 figures             figs        
#> 102 filesystem          fs          
#> 103 foreign key         fk          
#> 104 foreign keys        fks         
#> 105 formula             fm          
#> 106 formulas            fms         
#> 107 formulae            fms         
#> 108 frequent            freq        
#> 109 frequency           freq        
#> 110 frequencies         freqs       
#> 111 function            fn          
#> 112 functions           fns         
#> 113 generate            gen         
#> 114 generation          gen         
#> 115 generations         gens        
#> 116 identify            id          
#> 117 identifier          id          
#> 118 identifiers         ids         
#> 119 image               img         
#> 120 images              imgs        
#> 121 include             incl        
#> 122 inclusion           incl        
#> 123 index               i           
#> 124 indexes             ix          
#> 125 indices             ix          
#> 126 information         info        
#> 127 initialize          init        
#> 128 initialization      init        
#> 129 install             instl       
#> 130 installation        instl       
#> 131 integer             int         
#> 132 integers            ints        
#> 133 iterate             itr         
#> 134 iteration           itr         
#> 135 iterator            itr         
#> 136 iterations          itrs        
#> 137 iterators           itrs        
#> 138 label               lbl         
#> 139 labels              lbls        
#> 140 language            lang        
#> 141 languages           langs       
#> 142 left-hand side      lhs         
#> 143 level               lvl         
#> 144 levels              lvls        
#> 145 library             lib         
#> 146 libraries           libs        
#> 147 limit               lim         
#> 148 limits              lims        
#> 149 list                ls          
#> 150 logical             lgl         
#> 151 logicals            lgls        
#> 152 management          mgmt        
#> 153 Markdown            md          
#> 154 matrix              mat         
#> 155 matrices            mats        
#> 156 message             msg         
#> 157 messages            msgs        
#> 158 modify              mod         
#> 159 modification        mod         
#> 160 modifications       mods        
#> 161 not available       na          
#> 162 not a number        nan         
#> 163 number              nr          
#> 164 numbers             nrs         
#> 165 number of           n           
#> 166 numeric             num         
#> 167 numerics            nums        
#> 168 object              obj         
#> 169 objects             objs        
#> 170 operate             op          
#> 171 operation           op          
#> 172 operator            op          
#> 173 operations          ops         
#> 174 operators           ops         
#> 175 option              opt         
#> 176 options             opts        
#> 177 organize            org         
#> 178 organization        org         
#> 179 organizations       orgs        
#> 180 package             pkg         
#> 181 packages            pkgs        
#> 182 parameterize        param       
#> 183 parameter           param       
#> 184 parameters          params      
#> 185 position            pos         
#> 186 PostgreSQL          pg          
#> 187 predicate           pred        
#> 188 predicates          preds       
#> 189 prepare             prep        
#> 190 preparation         prep        
#> 191 preparations        preps       
#> 192 primary key         pk          
#> 193 primary keys        pks         
#> 194 proceed             prcd        
#> 195 procedure           prcd        
#> 196 procedures          prcds       
#> 197 projection          proj        
#> 198 project             proj        
#> 199 projections         projs       
#> 200 projects            projs       
#> 201 property            prop        
#> 202 properties          props       
#> 203 prototype           ptype       
#> 204 prototypes          ptypes      
#> 205 Quarto Markdown     qmd         
#> 206 question            qstn        
#> 207 questions           qstns       
#> 208 questionnaire       qstnr       
#> 209 questionnaires      qstnrs      
#> 210 R Markdown          rmd         
#> 211 refer               ref         
#> 212 reference           ref         
#> 213 references          refs        
#> 214 referendum          rfrnd       
#> 215 referendums         rfrnds      
#> 216 referenda           rfrnds      
#> 217 regular expression  regex       
#> 218 regular expressions regex       
#> 219 relative            rel         
#> 220 remove              rm          
#> 221 removal             rm          
#> 222 repository          repo        
#> 223 repositories        repos       
#> 224 request             req         
#> 225 requests            reqs        
#> 226 respond             resp        
#> 227 response            resp        
#> 228 responses           resps       
#> 229 right-hand side     rhs         
#> 230 roxygen2            roxy        
#> 231 separate            sep         
#> 232 separator           sep         
#> 233 separators          seps        
#> 234 sequence            seq         
#> 235 sequences           seqs        
#> 236 snippet             snip        
#> 237 snippets            snips       
#> 238 source              src         
#> 239 sources             srcs        
#> 240 specify             spec        
#> 241 specification       spec        
#> 242 specifications      specs       
#> 243 string              str         
#> 244 strings             strs        
#> 245 structure           struct      
#> 246 structures          structs     
#> 247 supplement          suppl       
#> 248 supplemental        suppl       
#> 249 supplementary       suppl       
#> 250 symbolize           sym         
#> 251 symbol              sym         
#> 252 symbols             syms        
#> 253 tabulate            tbl         
#> 254 table               tbl         
#> 255 tables              tbls        
#> 256 template            tpl         
#> 257 templates           tpls        
#> 258 temporary           tmp         
#> 259 user interface      ui          
#> 260 user experience     ux          
#> 261 value               val         
#> 262 values              vals        
#> 263 variable            var         
#> 264 variables           vars        
#> 265 vectorize           vctr        
#> 266 vector              vctr        
#> 267 vectors             vctrs       
#> 268 verbatim            verb        
#> 269 version             vrsn        
#> 270 versions            vrsns       
#> 271 working directory   wd