Markdown snippets

A tibble with all Markdown snippets together with their id. The latter can be fed to md_snip().

Usage

data_md_snips

Format

A tibble with the columns id and value.

Details

Currently, Markdown snippets with the following ids are available:

id

opt_global_max_cache_age

pkg_config

pkgdown_notice

rstudio_addin_hint

usage_notice

See also

Examples

pkgsnip::data_md_snips
#> # A tibble: 5 × 2
#>   id                       value                                                                                                                                
#>   <chr>                    <chr>                                                                                                                                
#> 1 opt_global_max_cache_age Maximal timespan to preserve the package's [pkgpins](https://pkgpins.rpkg.dev/) cache. Cache entries older than this will be deleted…
#> 2 pkg_config               Some of { pkg }'s functionality is controlled via package-specific global configuration which can either be set via [R options](http…
#> 3 pkgdown_notice           The documentation of this package is found [here]({ pkgdown::as_pkgdown()$meta$url }).                                               
#> 4 rstudio_addin_hint       This function is also registered as an [RStudio add-in](https://rstudio.github.io/rstudioaddins/), allowing RStudio users to assign …
#> 5 usage_notice             { pkg_intro_vignette <- stringr::str_replace_all(desc::desc_get_field('Package'), '.', '-'); if (fs::file_exists(fs::path('vignettes…