R condition messages

A tibble with all R condition messages included in this package, together with their id. The latter can be fed to msg().

Usage

data_msgs

Format

A tibble with the columns id and value.

Details

Currently, R condition messages with the following ids and arguments are available:

id	arguments

pkg_required	pkg

See also