Type hints

A tibble with all R object type hints together with their id. The latter can be fed to type().

Usage

data_types

Format

A tibble with the columns id, value, and is_scalar.

See also

Examples

pkgsnip::data_types
#> # A tibble: 11 × 3
#>    id      value                                               is_scalar
#>    <chr>   <chr>                                               <lgl>    
#>  1 chr     [`character({length})`][base::character]            TRUE     
#>  2 date    [`Date`][base::Date]                                FALSE    
#>  3 df      [`data.frame`][base::data.frame]                    FALSE    
#>  4 fct     [`factor`][base::factor]                            FALSE    
#>  5 int     [`integer({length})`][base::integer]                TRUE     
#>  6 lgl     [`logical({length})`][base::logical]                TRUE     
#>  7 list    [`list()`][base::list]                              FALSE    
#>  8 mat     [`matrix()`][base::matrix]                          FALSE    
#>  9 num     [`numeric({length})`][base::numeric]                TRUE     
#> 10 pg_conn [`PqConnection`][RPostgres::dbGetInfo_PqConnection] FALSE    
#> 11 raw     [`raw({length})`][base::raw]                        TRUE