Get predefined @description label

Returns a predefined label intended to be used as a function's description using roxygen2's @description tag.

Usage

description_lbl(id = roxy_lbls(type = "description")$id, ...)

Arguments

id

Label identifier. See roxy_lbls() for possible values.

...

Further named arguments passed on to roxy_lbl().

Value

A glue character scalar.

Details

A label can be inserted using inline R code as follows:

#' @description `r pkgsnip::description_lbl("ID")`

Note that the above only works in roxygen2 7.1.0+.

Currently, the following @description labels are available:

id	arguments

cli_markup_support

date

dates

datetime

datetimes

dyn_dots_support

end_date

eol

ggplot2_obj

glue_chr

gt_obj

max_cache_age

num_vrsn

opt_global_max_cache_age	pkg, global_max_cache_age

path

paths

pkg_config	pkg

plotly_obj

quiet

r_obj

resp_obj

start_date

strict_list

sym

tbl_data

tibble

tibble_cols	cols

tidy_select_support

use_cache

See also