List all available R Markdown file snippets

Lists all of the R Markdown snippets shipped with this package, together with the paths where they're located on the filesystem.

ls_file_snips()

Value

pkgsnip::return_label("data")

See also