Get predefined R condition message

Get predefined R condition message

message(name = messages()$name, ...)

msg(name = messages()$name, ...)

Arguments

name
The message name. See messages() for possible values.

...
Further named arguments used to tailor the message to your needs. Not all messages require additional arguments, see messages() for an overview.

Value

A character scalar.

See also

Examples

pkgsnip::message(name = "pkg_required",
                 pkg = "some_pkg")

#> To be able to use this function, the package 'some_pkg' is required but it is not installed. Please install it and then try again.