Get a table of all available R condition messages

Returns a tibble listing all R condition messages included in this package, together with their name which can be provided as message()'s name argument.

messages()

msgs()

Value

A tibble.

Details

Currently, R condition messages with the following names and arguments are available:

name	arguments

pkg_required	pkg

See also