Get predefined @param label

Returns a pre-defined label intended to be used to document a function parameter using roxygen2's @param tag.

Usage

param_label(name = roxy_labels(type = "param")$name, ...)

param_lbl(name = roxy_labels(type = "param")$name, ...)

Arguments

name

The label name. See roxy_labels() for possible values.

...

Further named arguments used to tailor the label to your needs. Not all labels require additional arguments, see roxy_labels() for an overview. If a required argument is not explicitly provided, it is searched for in the parent frames.

Value

A character vector.

Details

A label can be inserted using inline R code as follows:

#' @param PARAM_NAME `r pkgsnip::param_label("LABEL_NAME")`

Note that the above only works in roxygen2 7.1.0+.

Currently, the following parameter labels are available:

name	arguments

data

data_cols	pal::prose_ls(cols, wrap = '')`

date

dates

datetime

datetimes

gt_obj

path

paths

plotly_obj

r_obj

response

strict_list

symbol

tabular_data

version_nr

opt_global_max_cache_age	pkg, global_max_cache_age

cli_markup_support

dyn_dots_support

tidy_select_support

eol

quiet

start_date

end_date

use_cache

max_cache_age

See also