pkgsnip: Standardized Labels, Messages and (R) Markdown Snippets for Package Authors

pkgsnip ships ready-to-go (R) Markdown snippets and labels intended to be used in roxygen2 documentation or as knitr child documents. Also included are some predefined R condition messages and a table of abbreviations commonly used in R code. All of them are designed to reduce unnecessary repetition and standardize R-package documentation wherever possible.

See also

Author

Maintainer: Salim Brüggemann salim-b@pm.me (ORCID)