Get predefined return label

Returns a pre-defined label intended to be used to document function return values using roxygen2's @return tag.

return_label(name = roxy_labels(type = "return")$name, ...)

return_lbl(name = roxy_labels(type = "return")$name, ...)

Arguments

name
The label name. See roxy_labels() for possible values.

...
Further named arguments used to tailor the label to your needs. Not all labels require additional arguments, see roxy_labels() for an overview. If a required argument is not explicitly provided, it is searched for in the parent frames.

Value

A character vector.

Details

A label can be inserted using inline R code as follows:

#' @return version_nr `r pkgsnip::return_label("version_nr")`

Note that the above only works in roxygen2 7.1.0+.

Currently, the following return labels are available:

name	arguments

data

date

dates

datetime

datetimes

path

paths

r_obj

response

strict_list

symbol

tabular_data

version_nr

opt_max_cache_lifespan	this_pkg, max_cache_lifespan

See also