Get a table of all available roxygen2 tag labels

Returns a tibble listing all parameter labels included in this package, together with their name which can be provided as the name argument of param_label() or return_label().

roxy_labels(type = c("any", "param", "return", NA_character_))

roxy_lbls(type = c("any", "param", "return", NA_character_))

Arguments

type
The label type to return. A character scalar.

Value

A tibble.

Details

Currently, parameter labels with the following types and names are available:

type	name	arguments

NA	data

NA	date

NA	dates

NA	path

NA	paths

NA	r_object

NA	strict_list

NA	version_nr

NA	opt_max_cache_lifespan	pkg, max_cache_lifespan

param	dyn_dots_support

param	start_date

param	end_date

param	use_cache

param	cache_lifespan

See also