Get a table of all available roxygen2 tag labels

Returns a tibble listing all parameter labels included in this package, together with their id which can be provided as the id argument of roxy_lbl(), param_lbl(), return_lbl(), etc.

Usage

roxy_lbls(type = "default")

Arguments

type

The label type to return label values for. One of "default", "author", "concept", "description", "details", "field", "format", "keywords", "param", "references", "return", "section", "seealso", "slot", "source" or "title".

Value

A tibble.

Details

Currently, roxygen2 tag labels are available:

id	type	arguments

cli_markup_support	default

date	default

dates	default

datetime	default

datetimes	default

dbi_connection	default

df_or_tibble	default

dm	default

dyn_dots_support	default

end_date	default

eol	default

ggplot2_obj	default

glue_chr	default

gt_obj	default

max_cache_age	default

num_vrsn	default

opt_global_max_cache_age	default	pkg, global_max_cache_age

path	default

paths	default

pkg_config	default	pkg

pkg_config	title	pkg

plotly_obj	default

quiet	default

r_obj	default

resp_obj	default

show_progress	default

start_date	default

strict_list	default

sym	default

tbl_data	default

tibble	default

tibble_cols	default	cols

tibble_custom	default	custom

tidy_select_support	default

use_cache	default

See also