Get R Markdown snippet file path

Returns the file path to an R Markdown snippet shipped with this package.

snip_path(name = ls_file_snips()$name)

Arguments

name
The name of a snippet. Possible values include:

"coding-style-notice.Rmd"

"installation-notice_dev-version.Rmd"

"installation-notice_dev-version_gitlab.Rmd"

"rmd-package-notice.Rmd")`

Value

A path.

Details

The snippets can be used anywhere R Markdown input is supported. For example, you can use them as input to knitr's child document option:

```{r, child = pkgsnip::snip_path("installation-notice_dev-version_gitlab.Rmd")}
```

Or you can use them in roxygen2's @includeRmd tag:

#' @includeRmd `r pkgsnip::snip_path("installation-notice_dev-version_gitlab.Rmd")`

See also

Examples

pkgsnip::snip_path("coding-style-notice.Rmd")

#> /tmp/RtmpwVoFHs/temp_libpath102c78ebe552/pkgsnip/snippets/coding-style-notice.Rmd