Commonly used abbreviations in R code

Returns a tibble listing an opinionated set of abbreviations commonly used when writing R code. It can be used as a reference, for example, to check availability/consistency when considering using some abbreviation in a function or argument name.

Usage

abbreviations(expand = FALSE)

abbrs(expand = FALSE)

Arguments

expand

Whether to expand the full_expression column and return the data in long format. If FALSE, a "nested" list column full_expressions will be returned, meaning the values in column abbreviation will be unique.

Value

A tibble.

Details

To avoid the list getting excessively long and redundant, the main focus is on verbs and nouns. Adjectives and adverbs are only listed in column full_expression when there's not already a verb or noun with the same root word listed for a particular abbreviation. That means that e.g. the abbreviation "seq" could also stand for the full expressions "sequential" or "sequentially", despite the fact that it is not explicitly listed but merely the verb/noun "sequence" is.

Examples

pkgsnip::abbrs(expand = TRUE) |> print(n = Inf)
#> # A tibble: 177 × 2
#>     full_expression     abbreviation
#>     <chr>               <chr>       
#>   1 abbreviate          abbr        
#>   2 abbreviation        abbr        
#>   3 abbreviations       abbrs       
#>   4 absolute            abs         
#>   5 argument            arg         
#>   6 arguments           args        
#>   7 attribute           attr        
#>   8 attributes          attrs       
#>   9 authenticate        auth        
#>  10 authentication      auth        
#>  11 authentications     auths       
#>  12 background          bg          
#>  13 bibliography        bib         
#>  14 bibliographies      bibs        
#>  15 chapter             chpt        
#>  16 chapters            chpts       
#>  17 character           chr         
#>  18 characters          chrs        
#>  19 column              col         
#>  20 columns             cols        
#>  21 command             cmd         
#>  22 commands            cmds        
#>  23 combination         combo       
#>  24 combinations        combos      
#>  25 condition           cnd         
#>  26 conditions          cnds        
#>  27 configure           config      
#>  28 configuration       config      
#>  29 configurations      configs     
#>  30 database            db          
#>  31 dataframe           df          
#>  32 dataframes          dfs         
#>  33 dataframe column    dfc         
#>  34 dataframe row       dfr         
#>  35 define              def         
#>  36 definition          def         
#>  37 definitions         defs        
#>  38 depend              dep         
#>  39 dependency          dep         
#>  40 dependencies        deps        
#>  41 develop             dev         
#>  42 development         dev         
#>  43 developer           dev         
#>  44 developments        devs        
#>  45 developers          devs        
#>  46 dictionary          dict        
#>  47 dictionaries        dicts       
#>  48 differentiate       diff        
#>  49 difference          diff        
#>  50 differences         diffs       
#>  51 directory           dir         
#>  52 directories         dirs        
#>  53 distribution        distro      
#>  54 distributions       distros     
#>  55 document            doc         
#>  56 documents           docs        
#>  57 double              dbl         
#>  58 doubles             dbls        
#>  59 element             el          
#>  60 elements            els         
#>  61 environment         env         
#>  62 environments        envs        
#>  63 exclude             excl        
#>  64 exclusion           excl        
#>  65 expression          expr        
#>  66 expressions         exprs       
#>  67 factor              fct         
#>  68 factors             fcts        
#>  69 figure              fig         
#>  70 figures             figs        
#>  71 filesystem          fs          
#>  72 formula             fm          
#>  73 formulas            fms         
#>  74 formulae            fms         
#>  75 frequent            freq        
#>  76 frequency           freq        
#>  77 frequencies         freqs       
#>  78 function            fn          
#>  79 functions           fns         
#>  80 generate            gen         
#>  81 generation          gen         
#>  82 generations         gens        
#>  83 identify            id          
#>  84 identifier          id          
#>  85 identifiers         ids         
#>  86 include             incl        
#>  87 inclusion           incl        
#>  88 index               i           
#>  89 indexes             ix          
#>  90 indices             ix          
#>  91 information         info        
#>  92 initialize          init        
#>  93 initialization      init        
#>  94 integer             int         
#>  95 integers            ints        
#>  96 label               lbl         
#>  97 labels              lbls        
#>  98 language            lang        
#>  99 languages           langs       
#> 100 level               lvl         
#> 101 levels              lvls        
#> 102 list                ls          
#> 103 logical             lgl         
#> 104 logicals            lgls        
#> 105 management          mgmt        
#> 106 Markdown            md          
#> 107 message             msg         
#> 108 messages            msgs        
#> 109 modify              mod         
#> 110 modification        mod         
#> 111 modifications       mods        
#> 112 number              nr          
#> 113 numbers             nrs         
#> 114 number of           n           
#> 115 numeric             num         
#> 116 numerics            nums        
#> 117 object              obj         
#> 118 objects             objs        
#> 119 option              opt         
#> 120 options             opts        
#> 121 package             pkg         
#> 122 packages            pkgs        
#> 123 parameterize        param       
#> 124 parameter           param       
#> 125 parameters          params      
#> 126 proceed             prcd        
#> 127 procedure           prcd        
#> 128 procedures          prcds       
#> 129 prototype           ptype       
#> 130 prototypes          ptypes      
#> 131 Quarto Markdown     qmd         
#> 132 questionnaire       qstnr       
#> 133 questionnaires      qstnrs      
#> 134 R Markdown          rmd         
#> 135 refer               ref         
#> 136 reference           ref         
#> 137 references          refs        
#> 138 referendum          rfrnd       
#> 139 referendums         rfrnds      
#> 140 referenda           rfrnds      
#> 141 regular expression  regex       
#> 142 regular expressions regex       
#> 143 relative            rel         
#> 144 remove              rm          
#> 145 removal             rm          
#> 146 roxygen2            roxy        
#> 147 separate            sep         
#> 148 separator           sep         
#> 149 separators          seps        
#> 150 sequence            seq         
#> 151 sequences           seqs        
#> 152 specify             spec        
#> 153 specification       spec        
#> 154 specifications      specs       
#> 155 string              str         
#> 156 strings             strs        
#> 157 supplemental        suppl       
#> 158 supplementary       suppl       
#> 159 symbolize           sym         
#> 160 symbol              sym         
#> 161 symbols             syms        
#> 162 tabulate            tbl         
#> 163 table               tbl         
#> 164 tables              tbls        
#> 165 template            tpl         
#> 166 templates           tpls        
#> 167 temporary           tmp         
#> 168 value               val         
#> 169 values              vals        
#> 170 variable            var         
#> 171 variables           vars        
#> 172 vectorize           vctr        
#> 173 vector              vctr        
#> 174 vectors             vctrs       
#> 175 version             vrsn        
#> 176 versions            vrsns       
#> 177 working directory   wd