Commonly used abbreviations in R code

Returns a tibble listing an opinionated set of abbreviations commonly used when writing R code. It can be used as a reference, for example, to check availability/consistency when considering using some abbreviation in a function or argument name.

abbreviations(expand = FALSE)

abbrs(expand = FALSE)

Arguments

expand
Whether to expand the full_expression column and return the data in long format. If FALSE, a "nested" list column full_expressions will be returned, meaning the values in column abbreviation will be unique.

Value

A tibble.

Details

The main focus is on verbs and nouns. Adjectives and adverbs are only listed in column full_expression when there's not already a verb or noun with the same root word listed for a particular abbreviation. That means that e.g. the abbreviation "seq" could also stand for the full expression "sequential", despite the fact that it is not explicitly listed but merely the verb/noun "sequence" is.

Examples

pkgsnip::abbrs(expand = TRUE) |> print(n = Inf)
#> # A tibble: 165 × 2
#>     full_expression       abbreviation
#>     <chr>                 <chr>       
#>   1 abbreviate            abbr        
#>   2 abbreviation          abbr        
#>   3 abbreviations         abbrs       
#>   4 absolute              abs         
#>   5 argument              arg         
#>   6 arguments             args        
#>   7 attribute             attr        
#>   8 attributes            attrs       
#>   9 authenticate          auth        
#>  10 authentication        auth        
#>  11 authentications       auths       
#>  12 background            bg          
#>  13 bibliography          bib         
#>  14 bibliographies        bibs        
#>  15 chapter               chpt        
#>  16 chapters              chpts       
#>  17 character             chr         
#>  18 characters            chrs        
#>  19 column                col         
#>  20 columns               cols        
#>  21 command               cmd         
#>  22 commands              cmds        
#>  23 combination           combo       
#>  24 combinations          combos      
#>  25 condition             cnd         
#>  26 conditions            cnds        
#>  27 configure             config      
#>  28 configuration         config      
#>  29 configurations        configs     
#>  30 database              db          
#>  31 dataframe             df          
#>  32 dataframes            dfs         
#>  33 dataframe column      dfc         
#>  34 dataframe row         dfr         
#>  35 depend                dep         
#>  36 dependency            dep         
#>  37 dependencies          deps        
#>  38 develop               dev         
#>  39 development           dev         
#>  40 developer             dev         
#>  41 developments          devs        
#>  42 developers            devs        
#>  43 dict                  dictionary  
#>  44 dicts                 dictionaries
#>  45 differentiate         diff        
#>  46 difference            diff        
#>  47 differences           diffs       
#>  48 directory             dir         
#>  49 directories           dirs        
#>  50 distribution          distro      
#>  51 distributions         distros     
#>  52 document              doc         
#>  53 documents             docs        
#>  54 double                dbl         
#>  55 doubles               dbls        
#>  56 element               el          
#>  57 elements              els         
#>  58 environment           env         
#>  59 environments          envs        
#>  60 exclude               excl        
#>  61 exclusion             excl        
#>  62 expression            expr        
#>  63 expressions           exprs       
#>  64 factor                fct         
#>  65 factors               fcts        
#>  66 figure                fig         
#>  67 figures               figs        
#>  68 filesystem            fs          
#>  69 formula               fm          
#>  70 formulas              fms         
#>  71 formulae              fms         
#>  72 frequent              freq        
#>  73 frequency             freq        
#>  74 frequencies           freqs       
#>  75 function              fn          
#>  76 functions             fns         
#>  77 generate              gen         
#>  78 generation            gen         
#>  79 generations           gens        
#>  80 identify              id          
#>  81 identifier            id          
#>  82 identifiers           ids         
#>  83 include               incl        
#>  84 inclusion             incl        
#>  85 index                 i           
#>  86 indices/indexes       ix          
#>  87 information           info        
#>  88 initialize            init        
#>  89 initialization        init        
#>  90 integer               int         
#>  91 integers              ints        
#>  92 label                 lbl         
#>  93 labels                lbls        
#>  94 language              lang        
#>  95 languages             langs       
#>  96 level                 lvl         
#>  97 levels                lvls        
#>  98 list                  ls          
#>  99 logical               lgl         
#> 100 logicals              lgls        
#> 101 management            mgmt        
#> 102 Markdown              md          
#> 103 message               msg         
#> 104 messages              msgs        
#> 105 modify                mod         
#> 106 modification          mod         
#> 107 modifications         mods        
#> 108 number                nr          
#> 109 numbers               nrs         
#> 110 number of             n           
#> 111 numeric               num         
#> 112 numerics              nums        
#> 113 object                obj         
#> 114 objects               objs        
#> 115 option                opt         
#> 116 options               opts        
#> 117 package               pkg         
#> 118 packages              pkgs        
#> 119 parameterize          param       
#> 120 parameter             param       
#> 121 parameters            params      
#> 122 prototype             ptype       
#> 123 prototypes            ptypes      
#> 124 Quarto Markdown       qmd         
#> 125 R Markdown            rmd         
#> 126 refer                 ref         
#> 127 reference             ref         
#> 128 references            refs        
#> 129 referendum            rfrnd       
#> 130 referendums           rfrnds      
#> 131 referenda             rfrnds      
#> 132 regular expression(s) regex       
#> 133 relative              rel         
#> 134 remove                rm          
#> 135 removal               rm          
#> 136 roxygen2              roxy        
#> 137 separate              sep         
#> 138 separator             sep         
#> 139 separators            seps        
#> 140 sequence              seq         
#> 141 sequences             seqs        
#> 142 specify               spec        
#> 143 specification         spec        
#> 144 specifications        specs       
#> 145 string                str         
#> 146 strings               strs        
#> 147 symbolize             sym         
#> 148 symbol                sym         
#> 149 symbols               syms        
#> 150 tabulate              tbl         
#> 151 table                 tbl         
#> 152 tables                tbls        
#> 153 template              tpl         
#> 154 templates             tpls        
#> 155 temporary             tmp         
#> 156 value                 val         
#> 157 values                vals        
#> 158 variable              v           
#> 159 variables             vx          
#> 160 vectorize             vctr        
#> 161 vector                vctr        
#> 162 vectors               vctrs       
#> 163 version               vrsn        
#> 164 versions              vrsns       
#> 165 working directory     wd