Commonly used abbreviations in R code

Returns a tibble listing an opinionated set of abbreviations commonly used when writing R code. It can be used as a reference, for example, to check availability/consistency when considering using some abbreviation in a function or argument name.

Usage

abbrs(unnest = FALSE)

Arguments

unnest

Whether to unnest the full_expression column and return the data in long format. If FALSE, a "nested" list column full_expressions will be returned, meaning the values in column abbreviation will be unique.

Value

A tibble.

Details

To avoid the list getting excessively long and redundant, the main focus is on verbs and nouns. Adjectives and adverbs are only listed in column full_expression when there's not already a verb or noun with the same root word listed for a particular abbreviation. That means that e.g. the abbreviation "seq" could also stand for the full expressions "sequential" or "sequentially", despite the fact that it is not explicitly listed but merely the verb/noun "sequence" is.

Examples

pkgsnip::abbrs(unnest = TRUE) |> print(n = Inf)
#> # A tibble: 242 × 2
#>     full_expression     abbreviation
#>     <chr>               <chr>       
#>   1 abbreviate          abbr        
#>   2 abbreviation        abbr        
#>   3 abbreviations       abbrs       
#>   4 absolute            abs         
#>   5 argument            arg         
#>   6 arguments           args        
#>   7 attribute           attr        
#>   8 attributes          attrs       
#>   9 authenticate        auth        
#>  10 authentication      auth        
#>  11 authentications     auths       
#>  12 auxiliary           aux         
#>  13 background          bg          
#>  14 bibliography        bib         
#>  15 bibliographies      bibs        
#>  16 certificate         cert        
#>  17 certification       cert        
#>  18 certificates        certs       
#>  19 certifications      certs       
#>  20 chapter             chpt        
#>  21 chapters            chpts       
#>  22 character           chr         
#>  23 characters          chrs        
#>  24 column              col         
#>  25 columns             cols        
#>  26 command             cmd         
#>  27 commands            cmds        
#>  28 combination         combo       
#>  29 combinations        combos      
#>  30 condition           cnd         
#>  31 conditions          cnds        
#>  32 configure           config      
#>  33 configuration       config      
#>  34 configurations      configs     
#>  35 current             cur         
#>  36 database            db          
#>  37 dataframe           df          
#>  38 dataframes          dfs         
#>  39 dataframe column    dfc         
#>  40 dataframe row       dfr         
#>  41 define              def         
#>  42 definition          def         
#>  43 definitions         defs        
#>  44 delete              del         
#>  45 deletion            del         
#>  46 deletions           dels        
#>  47 depend              dep         
#>  48 dependency          dep         
#>  49 dependencies        deps        
#>  50 develop             dev         
#>  51 development         dev         
#>  52 developer           dev         
#>  53 developments        devs        
#>  54 developers          devs        
#>  55 dictionary          dict        
#>  56 dictionaries        dicts       
#>  57 differentiate       diff        
#>  58 difference          diff        
#>  59 differences         diffs       
#>  60 directory           dir         
#>  61 directories         dirs        
#>  62 distribution        distro      
#>  63 distributions       distros     
#>  64 document            doc         
#>  65 documents           docs        
#>  66 double              dbl         
#>  67 doubles             dbls        
#>  68 duplicate           dupl        
#>  69 duplication         dupl        
#>  70 duplicates          dupls       
#>  71 duplications        dupls       
#>  72 dynamic             dyn         
#>  73 element             el          
#>  74 elements            els         
#>  75 enumerate           enum        
#>  76 enumeration         enum        
#>  77 enumerations        enums       
#>  78 environment         env         
#>  79 environments        envs        
#>  80 evaluate            eval        
#>  81 evaluation          eval        
#>  82 evaluations         evals       
#>  83 exclude             excl        
#>  84 exclusion           excl        
#>  85 expression          expr        
#>  86 expressions         exprs       
#>  87 factor              fct         
#>  88 factors             fcts        
#>  89 figure              fig         
#>  90 figures             figs        
#>  91 filesystem          fs          
#>  92 foreign key         fk          
#>  93 foreign keys        fks         
#>  94 formula             fm          
#>  95 formulas            fms         
#>  96 formulae            fms         
#>  97 frequent            freq        
#>  98 frequency           freq        
#>  99 frequencies         freqs       
#> 100 function            fn          
#> 101 functions           fns         
#> 102 generate            gen         
#> 103 generation          gen         
#> 104 generations         gens        
#> 105 identify            id          
#> 106 identifier          id          
#> 107 identifiers         ids         
#> 108 image               img         
#> 109 images              imgs        
#> 110 include             incl        
#> 111 inclusion           incl        
#> 112 index               i           
#> 113 indexes             ix          
#> 114 indices             ix          
#> 115 information         info        
#> 116 initialize          init        
#> 117 initialization      init        
#> 118 integer             int         
#> 119 integers            ints        
#> 120 iterate             itr         
#> 121 iteration           itr         
#> 122 iterator            itr         
#> 123 iterations          itrs        
#> 124 iterators           itrs        
#> 125 label               lbl         
#> 126 labels              lbls        
#> 127 language            lang        
#> 128 languages           langs       
#> 129 left-hand side      lhs         
#> 130 level               lvl         
#> 131 levels              lvls        
#> 132 library             lib         
#> 133 libraries           libs        
#> 134 limit               lim         
#> 135 limits              lims        
#> 136 list                ls          
#> 137 logical             lgl         
#> 138 logicals            lgls        
#> 139 management          mgmt        
#> 140 Markdown            md          
#> 141 message             msg         
#> 142 messages            msgs        
#> 143 modify              mod         
#> 144 modification        mod         
#> 145 modifications       mods        
#> 146 NocoDB              ncdb        
#> 147 not available       na          
#> 148 not a number        nan         
#> 149 number              nr          
#> 150 numbers             nrs         
#> 151 number of           n           
#> 152 numeric             num         
#> 153 numerics            nums        
#> 154 object              obj         
#> 155 objects             objs        
#> 156 option              opt         
#> 157 options             opts        
#> 158 package             pkg         
#> 159 packages            pkgs        
#> 160 parameterize        param       
#> 161 parameter           param       
#> 162 parameters          params      
#> 163 position            pos         
#> 164 PostgreSQL          pg          
#> 165 predicate           pred        
#> 166 predicates          preds       
#> 167 preparation         prep        
#> 168 preparations        preps       
#> 169 primary key         pk          
#> 170 primary keys        pks         
#> 171 proceed             prcd        
#> 172 procedure           prcd        
#> 173 procedures          prcds       
#> 174 property            prop        
#> 175 properties          props       
#> 176 prototype           ptype       
#> 177 prototypes          ptypes      
#> 178 Quarto Markdown     qmd         
#> 179 question            qstn        
#> 180 questions           qstns       
#> 181 questionnaire       qstnr       
#> 182 questionnaires      qstnrs      
#> 183 R Markdown          rmd         
#> 184 refer               ref         
#> 185 reference           ref         
#> 186 references          refs        
#> 187 referendum          rfrnd       
#> 188 referendums         rfrnds      
#> 189 referenda           rfrnds      
#> 190 regular expression  regex       
#> 191 regular expressions regex       
#> 192 relative            rel         
#> 193 remove              rm          
#> 194 removal             rm          
#> 195 repository          repo        
#> 196 repositories        repos       
#> 197 request             req         
#> 198 requests            reqs        
#> 199 respond             resp        
#> 200 response            resp        
#> 201 responses           resps       
#> 202 right-hand side     rhs         
#> 203 roxygen2            roxy        
#> 204 separate            sep         
#> 205 separator           sep         
#> 206 separators          seps        
#> 207 sequence            seq         
#> 208 sequences           seqs        
#> 209 snippet             snip        
#> 210 snippets            snips       
#> 211 source              src         
#> 212 sources             srcs        
#> 213 specify             spec        
#> 214 specification       spec        
#> 215 specifications      specs       
#> 216 string              str         
#> 217 strings             strs        
#> 218 structure           struct      
#> 219 structures          structs     
#> 220 supplemental        suppl       
#> 221 supplementary       suppl       
#> 222 symbolize           sym         
#> 223 symbol              sym         
#> 224 symbols             syms        
#> 225 tabulate            tbl         
#> 226 table               tbl         
#> 227 tables              tbls        
#> 228 template            tpl         
#> 229 templates           tpls        
#> 230 temporary           tmp         
#> 231 user interface      ui          
#> 232 value               val         
#> 233 values              vals        
#> 234 variable            var         
#> 235 variables           vars        
#> 236 vectorize           vctr        
#> 237 vector              vctr        
#> 238 vectors             vctrs       
#> 239 verbatim            verb        
#> 240 version             vrsn        
#> 241 versions            vrsns       
#> 242 working directory   wd