Commonly used abbreviations in R code

Returns a tibble listing an opinionated set of abbreviations commonly used when writing R code. It can be used as a reference, for example, to check availability/consistency when considering using some abbreviation in a function or argument name.

Usage

abbrs(unnest = FALSE)

Arguments

unnest

Whether to unnest the full_expression column and return the data in long format. If FALSE, a "nested" list column full_expressions will be returned, meaning the values in column abbreviation will be unique.

Value

A tibble.

Details

To avoid the list getting excessively long and redundant, the main focus is on verbs and nouns. Adjectives and adverbs are only listed in column full_expression when there's not already a verb or noun with the same root word listed for a particular abbreviation. That means that e.g. the abbreviation "seq" could also stand for the full expressions "sequential" or "sequentially", despite the fact that it is not explicitly listed but merely the verb/noun "sequence" is.

Examples

pkgsnip::abbrs(unnest = TRUE) |> print(n = Inf)
#> # A tibble: 255 × 2
#>     full_expression     abbreviation
#>     <chr>               <chr>       
#>   1 abbreviate          abbr        
#>   2 abbreviation        abbr        
#>   3 abbreviations       abbrs       
#>   4 absolute            abs         
#>   5 argument            arg         
#>   6 arguments           args        
#>   7 attribute           attr        
#>   8 attributes          attrs       
#>   9 authenticate        auth        
#>  10 authentication      auth        
#>  11 authentications     auths       
#>  12 auxiliary           aux         
#>  13 background          bg          
#>  14 bibliography        bib         
#>  15 bibliographies      bibs        
#>  16 certificate         cert        
#>  17 certification       cert        
#>  18 certificates        certs       
#>  19 certifications      certs       
#>  20 chapter             chpt        
#>  21 chapters            chpts       
#>  22 character           chr         
#>  23 characters          chrs        
#>  24 column              col         
#>  25 columns             cols        
#>  26 command             cmd         
#>  27 commands            cmds        
#>  28 combination         combo       
#>  29 combinations        combos      
#>  30 condition           cnd         
#>  31 conditions          cnds        
#>  32 configure           config      
#>  33 configuration       config      
#>  34 configurations      configs     
#>  35 current             cur         
#>  36 database            db          
#>  37 dataframe           df          
#>  38 dataframes          dfs         
#>  39 dataframe column    dfc         
#>  40 dataframe row       dfr         
#>  41 define              def         
#>  42 definition          def         
#>  43 definitions         defs        
#>  44 delete              del         
#>  45 deletion            del         
#>  46 deletions           dels        
#>  47 depend              dep         
#>  48 dependency          dep         
#>  49 dependencies        deps        
#>  50 develop             dev         
#>  51 development         dev         
#>  52 developer           dev         
#>  53 developments        devs        
#>  54 developers          devs        
#>  55 dictionary          dict        
#>  56 dictionaries        dicts       
#>  57 differentiate       diff        
#>  58 difference          diff        
#>  59 differences         diffs       
#>  60 directory           dir         
#>  61 directories         dirs        
#>  62 distribution        distro      
#>  63 distributions       distros     
#>  64 document            doc         
#>  65 documents           docs        
#>  66 double              dbl         
#>  67 doubles             dbls        
#>  68 duplicate           dupl        
#>  69 duplication         dupl        
#>  70 duplicates          dupls       
#>  71 duplications        dupls       
#>  72 dynamic             dyn         
#>  73 element             el          
#>  74 elements            els         
#>  75 enumerate           enum        
#>  76 enumeration         enum        
#>  77 enumerations        enums       
#>  78 environment         env         
#>  79 environments        envs        
#>  80 evaluate            eval        
#>  81 evaluation          eval        
#>  82 evaluations         evals       
#>  83 execute             exec        
#>  84 execution           exec        
#>  85 executions          execs       
#>  86 exclude             excl        
#>  87 exclusion           excl        
#>  88 expression          expr        
#>  89 expressions         exprs       
#>  90 extend              ext         
#>  91 extension           ext         
#>  92 extensions          exts        
#>  93 factor              fct         
#>  94 factors             fcts        
#>  95 figure              fig         
#>  96 figures             figs        
#>  97 filesystem          fs          
#>  98 foreign key         fk          
#>  99 foreign keys        fks         
#> 100 formula             fm          
#> 101 formulas            fms         
#> 102 formulae            fms         
#> 103 frequent            freq        
#> 104 frequency           freq        
#> 105 frequencies         freqs       
#> 106 function            fn          
#> 107 functions           fns         
#> 108 generate            gen         
#> 109 generation          gen         
#> 110 generations         gens        
#> 111 identify            id          
#> 112 identifier          id          
#> 113 identifiers         ids         
#> 114 image               img         
#> 115 images              imgs        
#> 116 include             incl        
#> 117 inclusion           incl        
#> 118 index               i           
#> 119 indexes             ix          
#> 120 indices             ix          
#> 121 information         info        
#> 122 initialize          init        
#> 123 initialization      init        
#> 124 install             instl       
#> 125 installation        instl       
#> 126 integer             int         
#> 127 integers            ints        
#> 128 iterate             itr         
#> 129 iteration           itr         
#> 130 iterator            itr         
#> 131 iterations          itrs        
#> 132 iterators           itrs        
#> 133 label               lbl         
#> 134 labels              lbls        
#> 135 language            lang        
#> 136 languages           langs       
#> 137 left-hand side      lhs         
#> 138 level               lvl         
#> 139 levels              lvls        
#> 140 library             lib         
#> 141 libraries           libs        
#> 142 limit               lim         
#> 143 limits              lims        
#> 144 list                ls          
#> 145 logical             lgl         
#> 146 logicals            lgls        
#> 147 management          mgmt        
#> 148 Markdown            md          
#> 149 message             msg         
#> 150 messages            msgs        
#> 151 modify              mod         
#> 152 modification        mod         
#> 153 modifications       mods        
#> 154 not available       na          
#> 155 not a number        nan         
#> 156 number              nr          
#> 157 numbers             nrs         
#> 158 number of           n           
#> 159 numeric             num         
#> 160 numerics            nums        
#> 161 object              obj         
#> 162 objects             objs        
#> 163 operate             op          
#> 164 operation           op          
#> 165 operator            op          
#> 166 operations          ops         
#> 167 operators           ops         
#> 168 option              opt         
#> 169 options             opts        
#> 170 package             pkg         
#> 171 packages            pkgs        
#> 172 parameterize        param       
#> 173 parameter           param       
#> 174 parameters          params      
#> 175 position            pos         
#> 176 PostgreSQL          pg          
#> 177 predicate           pred        
#> 178 predicates          preds       
#> 179 preparation         prep        
#> 180 preparations        preps       
#> 181 primary key         pk          
#> 182 primary keys        pks         
#> 183 proceed             prcd        
#> 184 procedure           prcd        
#> 185 procedures          prcds       
#> 186 property            prop        
#> 187 properties          props       
#> 188 prototype           ptype       
#> 189 prototypes          ptypes      
#> 190 Quarto Markdown     qmd         
#> 191 question            qstn        
#> 192 questions           qstns       
#> 193 questionnaire       qstnr       
#> 194 questionnaires      qstnrs      
#> 195 R Markdown          rmd         
#> 196 refer               ref         
#> 197 reference           ref         
#> 198 references          refs        
#> 199 referendum          rfrnd       
#> 200 referendums         rfrnds      
#> 201 referenda           rfrnds      
#> 202 regular expression  regex       
#> 203 regular expressions regex       
#> 204 relative            rel         
#> 205 remove              rm          
#> 206 removal             rm          
#> 207 repository          repo        
#> 208 repositories        repos       
#> 209 request             req         
#> 210 requests            reqs        
#> 211 respond             resp        
#> 212 response            resp        
#> 213 responses           resps       
#> 214 right-hand side     rhs         
#> 215 roxygen2            roxy        
#> 216 separate            sep         
#> 217 separator           sep         
#> 218 separators          seps        
#> 219 sequence            seq         
#> 220 sequences           seqs        
#> 221 snippet             snip        
#> 222 snippets            snips       
#> 223 source              src         
#> 224 sources             srcs        
#> 225 specify             spec        
#> 226 specification       spec        
#> 227 specifications      specs       
#> 228 string              str         
#> 229 strings             strs        
#> 230 structure           struct      
#> 231 structures          structs     
#> 232 supplemental        suppl       
#> 233 supplementary       suppl       
#> 234 symbolize           sym         
#> 235 symbol              sym         
#> 236 symbols             syms        
#> 237 tabulate            tbl         
#> 238 table               tbl         
#> 239 tables              tbls        
#> 240 template            tpl         
#> 241 templates           tpls        
#> 242 temporary           tmp         
#> 243 user interface      ui          
#> 244 user experience     ux          
#> 245 value               val         
#> 246 values              vals        
#> 247 variable            var         
#> 248 variables           vars        
#> 249 vectorize           vctr        
#> 250 vector              vctr        
#> 251 vectors             vctrs       
#> 252 verbatim            verb        
#> 253 version             vrsn        
#> 254 versions            vrsns       
#> 255 working directory   wd