Commonly used abbreviations in R code

Returns a tibble listing an opinionated set of abbreviations commonly used when writing R code. It can be used as a reference, for example, to check availability/consistency when considering using some abbreviation in a function or argument name.

Usage

abbrs(unnest = FALSE)

Arguments

unnest

Whether to unnest the full_expression column and return the data in long format. If FALSE, a "nested" list column full_expressions will be returned, meaning the values in column abbreviation will be unique.

Value

A tibble.

Details

To avoid the list getting excessively long and redundant, the main focus is on verbs and nouns. Adjectives and adverbs are only listed in column full_expression when there's not already a verb or noun with the same root word listed for a particular abbreviation. That means that e.g. the abbreviation "seq" could also stand for the full expressions "sequential" or "sequentially", despite the fact that it is not explicitly listed but merely the verb/noun "sequence" is.

Examples

pkgsnip::abbrs(unnest = TRUE) |> print(n = Inf)
#> # A tibble: 212 × 2
#>     full_expression     abbreviation
#>     <chr>               <chr>       
#>   1 abbreviate          abbr        
#>   2 abbreviation        abbr        
#>   3 abbreviations       abbrs       
#>   4 absolute            abs         
#>   5 argument            arg         
#>   6 arguments           args        
#>   7 attribute           attr        
#>   8 attributes          attrs       
#>   9 authenticate        auth        
#>  10 authentication      auth        
#>  11 authentications     auths       
#>  12 background          bg          
#>  13 bibliography        bib         
#>  14 bibliographies      bibs        
#>  15 chapter             chpt        
#>  16 chapters            chpts       
#>  17 character           chr         
#>  18 characters          chrs        
#>  19 column              col         
#>  20 columns             cols        
#>  21 command             cmd         
#>  22 commands            cmds        
#>  23 combination         combo       
#>  24 combinations        combos      
#>  25 condition           cnd         
#>  26 conditions          cnds        
#>  27 configure           config      
#>  28 configuration       config      
#>  29 configurations      configs     
#>  30 current             cur         
#>  31 database            db          
#>  32 dataframe           df          
#>  33 dataframes          dfs         
#>  34 dataframe column    dfc         
#>  35 dataframe row       dfr         
#>  36 define              def         
#>  37 definition          def         
#>  38 definitions         defs        
#>  39 depend              dep         
#>  40 dependency          dep         
#>  41 dependencies        deps        
#>  42 develop             dev         
#>  43 development         dev         
#>  44 developer           dev         
#>  45 developments        devs        
#>  46 developers          devs        
#>  47 dictionary          dict        
#>  48 dictionaries        dicts       
#>  49 differentiate       diff        
#>  50 difference          diff        
#>  51 differences         diffs       
#>  52 directory           dir         
#>  53 directories         dirs        
#>  54 distribution        distro      
#>  55 distributions       distros     
#>  56 document            doc         
#>  57 documents           docs        
#>  58 double              dbl         
#>  59 doubles             dbls        
#>  60 element             el          
#>  61 elements            els         
#>  62 environment         env         
#>  63 environments        envs        
#>  64 evaluate            eval        
#>  65 evaluation          eval        
#>  66 evaluations         evals       
#>  67 exclude             excl        
#>  68 exclusion           excl        
#>  69 expression          expr        
#>  70 expressions         exprs       
#>  71 factor              fct         
#>  72 factors             fcts        
#>  73 figure              fig         
#>  74 figures             figs        
#>  75 filesystem          fs          
#>  76 formula             fm          
#>  77 formulas            fms         
#>  78 formulae            fms         
#>  79 frequent            freq        
#>  80 frequency           freq        
#>  81 frequencies         freqs       
#>  82 function            fn          
#>  83 functions           fns         
#>  84 generate            gen         
#>  85 generation          gen         
#>  86 generations         gens        
#>  87 identify            id          
#>  88 identifier          id          
#>  89 identifiers         ids         
#>  90 image               img         
#>  91 images              imgs        
#>  92 include             incl        
#>  93 inclusion           incl        
#>  94 index               i           
#>  95 indexes             ix          
#>  96 indices             ix          
#>  97 information         info        
#>  98 initialize          init        
#>  99 initialization      init        
#> 100 integer             int         
#> 101 integers            ints        
#> 102 label               lbl         
#> 103 labels              lbls        
#> 104 language            lang        
#> 105 languages           langs       
#> 106 left-hand side      lhs         
#> 107 level               lvl         
#> 108 levels              lvls        
#> 109 library             lib         
#> 110 libraries           libs        
#> 111 limit               lim         
#> 112 limits              lims        
#> 113 list                ls          
#> 114 logical             lgl         
#> 115 logicals            lgls        
#> 116 management          mgmt        
#> 117 Markdown            md          
#> 118 message             msg         
#> 119 messages            msgs        
#> 120 modify              mod         
#> 121 modification        mod         
#> 122 modifications       mods        
#> 123 not available       na          
#> 124 not a number        nan         
#> 125 number              nr          
#> 126 numbers             nrs         
#> 127 number of           n           
#> 128 numeric             num         
#> 129 numerics            nums        
#> 130 object              obj         
#> 131 objects             objs        
#> 132 option              opt         
#> 133 options             opts        
#> 134 package             pkg         
#> 135 packages            pkgs        
#> 136 parameterize        param       
#> 137 parameter           param       
#> 138 parameters          params      
#> 139 position            pos         
#> 140 predicate           pred        
#> 141 predicates          preds       
#> 142 preparation         prep        
#> 143 preparations        preps       
#> 144 proceed             prcd        
#> 145 procedure           prcd        
#> 146 procedures          prcds       
#> 147 property            prop        
#> 148 properties          props       
#> 149 prototype           ptype       
#> 150 prototypes          ptypes      
#> 151 Quarto Markdown     qmd         
#> 152 questionnaire       qstnr       
#> 153 questionnaires      qstnrs      
#> 154 R Markdown          rmd         
#> 155 refer               ref         
#> 156 reference           ref         
#> 157 references          refs        
#> 158 referendum          rfrnd       
#> 159 referendums         rfrnds      
#> 160 referenda           rfrnds      
#> 161 regular expression  regex       
#> 162 regular expressions regex       
#> 163 relative            rel         
#> 164 remove              rm          
#> 165 removal             rm          
#> 166 repository          repo        
#> 167 repositories        repos       
#> 168 request             req         
#> 169 requests            reqs        
#> 170 respond             resp        
#> 171 response            resp        
#> 172 responses           resps       
#> 173 right-hand side     rhs         
#> 174 roxygen2            roxy        
#> 175 separate            sep         
#> 176 separator           sep         
#> 177 separators          seps        
#> 178 sequence            seq         
#> 179 sequences           seqs        
#> 180 snippet             snip        
#> 181 snippets            snips       
#> 182 source              src         
#> 183 sources             srcs        
#> 184 specify             spec        
#> 185 specification       spec        
#> 186 specifications      specs       
#> 187 string              str         
#> 188 strings             strs        
#> 189 structure           struct      
#> 190 structures          structs     
#> 191 supplemental        suppl       
#> 192 supplementary       suppl       
#> 193 symbolize           sym         
#> 194 symbol              sym         
#> 195 symbols             syms        
#> 196 tabulate            tbl         
#> 197 table               tbl         
#> 198 tables              tbls        
#> 199 template            tpl         
#> 200 templates           tpls        
#> 201 temporary           tmp         
#> 202 value               val         
#> 203 values              vals        
#> 204 variable            var         
#> 205 variables           vars        
#> 206 vectorize           vctr        
#> 207 vector              vctr        
#> 208 vectors             vctrs       
#> 209 verbatim            verb        
#> 210 version             vrsn        
#> 211 versions            vrsns       
#> 212 working directory   wd