Get a table of all available roxygen2 tag labels

Returns a tibble listing all parameter labels included in this package, together with their name which can be provided as the name argument of param_label() or return_label().

roxy_labels(type = c("any", "param", "return", NA_character_))

roxy_lbls(type = c("any", "param", "return", NA_character_))

Arguments

type

The label type to return. A character scalar.

Value

A tibble.

Details

Currently, parameter labels with the following types and names are available:

type	name	arguments

NA	data

NA	data_cols	pal::prose_ls(cols, wrap = '')`

NA	date

NA	dates

NA	datetime

NA	datetimes

NA	gt_obj

NA	path

NA	paths

NA	plotly_obj

NA	r_obj

NA	response

NA	strict_list

NA	symbol

NA	tabular_data

NA	version_nr

NA	opt_max_cache_lifespan	this_pkg, max_cache_lifespan

param	cli_markup_support

param	dyn_dots_support

param	tidy_select_support

param	eol

param	quiet

param	start_date

param	end_date

param	use_cache

param	cache_lifespan

See also